move client id to global constant file
